backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
python-dotenv==1.0.1
neo4j==5.20.0
